Bug fix (Linux) for Live mode - updating SetupCameraForSeqAcquisition API to use a long instead of int. Reviewed by M.Milford.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@12914 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/AndorSDK3/AndorSDK3.cpp
@@ -968,7 +968,7 @@ bool CAndorSDK3Camera::CleanUpDeviceCircularBuffer()
    return true;
 }
 
-int CAndorSDK3Camera::SetupCameraForSeqAcquisition(int numImages)
+int CAndorSDK3Camera::SetupCameraForSeqAcquisition(long numImages)
 {
    int retCode = DEVICE_OK;
    bool b_memOkRet = false;
